requirements.txt
python-dotenv~=0.13.0
requests~=2.23.0
numpy~=1.18.2
pandas~=1.0.3
matplotlib~=3.2.1
seaborn
datetime~=4.3
flask>=1.0.0
scikit-learn
bayesian-optimization==0.6
kiteconnect~=3.8.2
dash

plotly~=4.7.0
APScheduler~=3.6.3
config~=0.5.0.post0
pytz~=2019.3
TA-Lib
colored
aiohttp
lxml~=4.5.1
multitasking~=0.0.7
h5py
interface
